third_party/blink/web_tests/virtual/xslt-disabled/fast/xsl/xslt/xslt-text-quirk-expected.txt

CONSOLE WARNING: XSLTProcessor and XSLT Processing Instructions have been removed in this browser. See https://chromestatus.com/feature/4709671889534976.
SOURCE XML: &lt;&lt;&lt;&amp;тест&amp;&gt;&gt;&gt;
XML viewer in iframe: true
XML viewer shows source XML: true
